web-service/requirements.txt
pika==1.3.2
Flask==3.1.0
requests==2.32.3
python-dotenv==1.1.0
msal==1.32.0
gunicorn
colorlog==6.9.0